Refactoring package steps. add 2 new class for steps: PageBlockSteps and RoundUpSteps
